refactor: Extract 기능, 컨벤션에 맞게 변경 (#60)
