+++ src/main/kotlin/app/askresume/domain/extract/service/ExtractService.kt
@@ -1,0 +1,30 @@
+package app.askresume.domain.extract.service
+
+import app.askresume.api.extract.vo.ExtractedTextResponse
+import app.askresume.global.util.LoggerUtil.log
+import org.apache.pdfbox.pdmodel.PDDocument
+import org.apache.pdfbox.text.PDFTextStripper
+import org.springframework.stereotype.Service
+import org.springframework.web.multipart.MultipartFile
+
+
+@Service
+class ExtractService {
+
+    val log = log()
+
+    fun pdfToText(file: MultipartFile): ExtractedTextResponse {
+        var text: String
+        file.inputStream.use { inputStream ->
+            val document = PDDocument.load(inputStream)
+            val pdfTextStripper = PDFTextStripper()
+
+            text = pdfTextStripper.getText(document)
+            document.close()
+        }
+
+        return ExtractedTextResponse(text)
+    }
+
+}
+
