+++ src/main/resources/message/messages_en.properties
@@ -16,7 +16,6 @@ signup.password.inconsistency=The password and password confirmation values do n
 # \uC774\uB825\uC11C
 invalid.difficulty.type=Invalid difficulty type.
 # \uD30C\uC77C \uC785\uCD9C\uB825
-not.permitted.content.type=The specified CONTENT TYPE is not permitted.
 text.extraction.failed=An error occurred during text extraction.
 # JOB
 job.not.exists=The specified job does not exist.
@@ -27,6 +26,9 @@ json.parsing.failed=An error occurred during JSON parsing.
 omitting.required.values=A required value has been omitted.
 
 # ------------------------------------------신규------------------------------------------------
+
+# FILE
+not.permitted.content.type=The specified CONTENT TYPE is not permitted.
 
 # Difficulty
 difficulty.not.exists=Difficulty level that does not exist.
