+++ src/main/resources/message/messages_ko.properties
@@ -16,7 +16,6 @@ signup.password.inconsistency=\uC554\uD638 \uBC0F \uC554\uD638 \uD655\uC778 \uAC
 # \uC774\uB825\uC11C
 invalid.difficulty.type=\uB09C\uC774\uB3C4\uAC00 \uC798\uBABB\uB418\uC5C8\uC2B5\uB2C8\uB2E4.
 # \uD30C\uC77C \uC785\uCD9C\uB825
-not.permitted.content.type=\uC9C0\uC815\uB41C CONTENT TYPE\uC774 \uD5C8\uC6A9\uB418\uC9C0 \uC54A\uC2B5\uB2C8\uB2E4.
 text.extraction.failed=\uD14D\uC2A4\uD2B8\uB97C \uCD94\uCD9C\uD558\uB294 \uB3D9\uC548 \uC624\uB958\uAC00 \uBC1C\uC0DD\uD588\uC2B5\uB2C8\uB2E4.
 # JOB
 job.not.exists=\uC9C0\uC815\uD55C \uC791\uC5C5\uC774 \uC874\uC7AC\uD558\uC9C0 \uC54A\uC2B5\uB2C8\uB2E4.
@@ -27,6 +26,9 @@ json.parsing.failed=JSON \uAD6C\uBB38 \uBD84\uC11D \uC911\uC5D0 \uC624\uB958\uAC
 omitting.required.values=\uD544\uC218\uAC12\uC774 \uC0DD\uB7B5\uB418\uC5C8\uC2B5\uB2C8\uB2E4.
 
 # ------------------------------------------신규------------------------------------------------
+
+# FILE
+not.permitted.content.type=지정된 CONTENT TYPE은 허용되지 않습니다.
 
 # Difficulty
 difficulty.not.exists=존재하지 않는 난이도입니다.
